pyproject.toml
[tool.poetry]
name = "inventory"
version = "0.1.0"
description = ""
authors = []
readme = "readme.md"
packages = [{include = "inventory"}]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.2"
peewee = "^3.16.0"
pydantic = "^1.10.7"
dependency-injector = "^4.41.0"
loguru = "^0.6.0"
loguru-logging-intercept = "^0.1.1"
aenum = "^3.1.12"
psycopg2-binary = "^2.9.5"
nest-asyncio = "^1.5.6"

[tool.poetry.group.dev.dependencies]
peewee-migrations = "^0.3.32"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
